+++ src/product-service/handlers/setProduct.ts
@@ -2,7 +2,7 @@
 import { uuid } from 'uuidv4';
 import 'source-map-support/register';
 import { Client } from 'pg';
-import { dbOptions } from '../../utils/dbOptions';
+import { dbOptions } from '../dbOptions';
 import { corsHeaders } from '../../utils/corsHeaders';
 
 export const setProduct = async event => {
